Refactored code.
