+++ live-site/src/pages/index.tsx
@@ -13,7 +13,7 @@ import ComingUpSection from '../../components/coming-up/ComingUp';
 import useMatchMedia from 'react-use-match-media';
 import MentorsSection from '../../components/mentors-section/MentorsSection';
 import { StyledPageContainer } from '../../../shared-ui/styled-components/Background.styles';
-import ToggleMode from '../../components/toggle-mode/ToggleMode';
+import ToggleMode from '../../../shared-ui/components/toggle-mode/ToggleMode';
 import EventScheduleSection from '../../components/event-schedule-section/EventScheduleSection';
 
 const handleMode = (): boolean => {
@@ -50,7 +50,7 @@ const IndexPage: React.FC = () => {
   return (
     <StyledPageContainer className={getBackgroundClassName()}>
       <Header tabs={liveSiteTabInfo} isDay={true} />
-      <ToggleMode isDay={isDay} setIsDay={setIsDay} />
+      <ToggleMode isDay={isDay} setIsDay={setIsDay} location={'live-site'} />
       {isDesktop && new Date() > new Date('2023-02-10T17:00:00-05:00') && (
         <HackingRemaining />
       )}
